+++ resource/resource_suite_test.go
@@ -1,8 +1,6 @@
 package resource
 
 import (
-	"log"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"gopkg.in/jose.v1/jwt"
@@ -26,12 +24,12 @@ func TestMongodb(t *testing.T) {
 var _ = AfterSuite(func() {
 
 	if db != nil {
-		log.Println("Closing connection to MongoDB")
+		log.Info("Closing connection to MongoDB")
 		db.Close()
 	}
 
 	if pool != nil {
-		log.Println("Purging containers")
+		log.Info("Purging containers")
 		if err := pool.Purge(containerResource); err != nil {
 			log.Fatalf("Could not purge resource: %s", err)
 		}