+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Prometheus metrics collection via `/metrics` endpoint
+
 ### Changed
 
 - Add access control to trip resource
 - New trips need to have a valid car share on creation
+- Switch to Gin Framework
+- Make firebase project configurable
 
 ## [0.3.3] - 2017-03-27